+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.1] - 2023-12-14
+
+### Updated
+- Updated `useCloseWithEscapeKey` hook:
+ - Made `keyDownTargetRef` optional
+ - If `keyDownTargetRef` is not used, use the document element
+ - Changed order of props now that `keyDownTargetRef` is optional
+ - Updated previous instances of `useCloseWithEscapeKey` to use new function signature
+
 ## [2.6.0] - 2023-12-13
 
 ### Added